Guard callback destructors during Python shutdown
