+++ source/isaaclab/isaaclab/sensors/sensor_base.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 import inspect
 import logging
 import re
+import sys
 import weakref
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
@@ -77,10 +78,11 @@ class SensorBase(ABC):
         # set initial state of debug visualization
         self.set_debug_vis(self.cfg.debug_vis)
 
-    def __del__(self):
+    def __del__(self, _sys=sys):
         """Unsubscribe from the callbacks."""
         # clear physics events handles
-        self._clear_callbacks()
+        if not _sys.is_finalizing() and _sys.meta_path is not None:
+            self._clear_callbacks()
 
     """
     Properties
